refactor(core): extract screenshot location + security into maestro-screenshot-file.js
Carve the file-location concern (platform glob pattern, manualScreenshotWalk
fallback, multi-match mtime selection, realpath + session-root prefix check)
out of handleMaestroScreenshot into locateScreenshot(). The handler now passes
the shape-validated filePath and gets back a canonicalized absolute path or a
404. Move the semgrep path-traversal suppression to the new module (it now
owns the path.join sinks). istanbul-ignore annotations travel verbatim.
No behavior change.
