+++ .semgrepignore
@@ -20,17 +20,17 @@ packages/core/src/lock.js
 # every join. Suppress at the file level with this rationale.
 packages/core/src/archive.js
 
-# maestro-screenshot.js path-traversal guards live at the top of the
-# handleMaestroScreenshot handler: `name` and `sessionId` are both validated
-# against /^[a-zA-Z0-9_-]+$/ (SAFE_ID) BEFORE any path.join() runs.
-# Traversal sequences (`..`, `/`, `\0`) are rejected with 400 there. The
-# fallback walker (manualScreenshotWalk) is also depth-capped at 15 levels.
-# semgrep's path-join-resolve-traversal rule cannot follow the regex
-# validation chain across the function body, so it flags the joins inside
-# manualScreenshotWalk. Inline `// nosemgrep` directives (preceding and
-# same-line) were not honored by the semgrep CI version in use — suppress at
-# the file level with this rationale.
-packages/core/src/maestro-screenshot.js
+# maestro-screenshot-file.js owns the screenshot-location path joins. The
+# handleMaestroScreenshot handler validates `name` and `sessionId` against
+# /^[a-zA-Z0-9_-]+$/ (SAFE_ID) and rejects traversal sequences (`..`, `/`,
+# `\0`) with 400 BEFORE calling locateScreenshot(); the fallback walker
+# (manualScreenshotWalk) is additionally depth-capped at 15 levels. semgrep's
+# path-join-resolve-traversal rule cannot follow that validation chain across
+# the module boundary, so it flags the joins inside manualScreenshotWalk.
+# Inline `// nosemgrep` directives (preceding and same-line) were not honored
+# by the semgrep CI version in use — suppress at the file level with this
+# rationale.
+packages/core/src/maestro-screenshot-file.js
 
 # Test files load fixtures from hardcoded subpaths under
 # test/fixtures/maestro-{hierarchy,ios-hierarchy}. The fixture filenames
